pyproject.toml
[project]
name = "coffee-tracker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.84.0",
    "flask>=3.1.3",
    "pillow>=12.1.1",
    "pillow-heif>=1.3.0",
    "python-dotenv>=1.2.2",
]
